more reasonable files and tests.
